+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Support empty structs with `CborSerialize` derive macro
+- Support CBOR decoding maps and arrays of indefinite length
 
 ## 8.0.0-alpha.2 (2025-07-14)